Move ParseTimeSpan to seperate utility class
